nels-oauth2/requirements.txt
appdirs==1.4.3
backports-abc==0.5
certifi==2017.1.23
packaging==16.8
pycrypto==2.6.1
pyparsing==2.2.0
python-oauth2==1.0.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.2
